Доделал фильтр Собела (впрочем, для распараллеливания -- его придётся чуть переделать); Результат уже выводится в GUI;

+++ headers/RGBCell.h
@@ -20,7 +20,7 @@ public:
     void setGreenValue(uint8_t value);
     void setBlueValue(uint8_t value);
 
-    uint32_t getLuminosity();
+    uint8_t getLuminosity();
 };
 
 #endif //RGBCELL_H
